requirements.txt
fastapi
uvicorn
psycopg2
SQLAlchemy
pyjwt
python-jose[cryptography]
python-multipart
bcrypt
pytest
